Refactored input sub-folders
