+++ front/src/modules/activities/timeline/components/TimelineActivityTitle.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import {
   Checkbox,
   CheckboxShape,
-} from '@/ui/input/checkbox/components/Checkbox';
+} from '@/ui/input/components/Checkbox';
 import { OverflowingTextWithTooltip } from '@/ui/tooltip/OverflowingTextWithTooltip';
 import { ActivityType } from '~/generated/graphql';
