+++ front/src/modules/settings/profile/components/ToggleField.tsx
@@ -1,7 +1,7 @@
 import { useRecoilValue } from 'recoil';
 
 import { currentUserState } from '@/auth/states/currentUserState';
-import { Toggle } from '@/ui/input/toggle/components/Toggle';
+import { Toggle } from '@/ui/input/components/Toggle';
 import { useSnackBar } from '@/ui/snack-bar/hooks/useSnackBar';
 import { useUpdateAllowImpersonationMutation } from '~/generated/graphql';
