+++ front/src/modules/spreadsheet-import/steps/components/SelectSheetStep/SelectSheetStep.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 
 import { ContinueButton } from '@/spreadsheet-import/components/ContinueButton';
 import { Heading } from '@/spreadsheet-import/components/Heading';
-import { Radio } from '@/ui/input/radio/components/Radio';
-import { RadioGroup } from '@/ui/input/radio/components/RadioGroup';
+import { Radio } from '@/ui/input/components/Radio';
+import { RadioGroup } from '@/ui/input/components/RadioGroup';
 import { Modal } from '@/ui/modal/components/Modal';
 
 const StyledContent = styled(Modal.Content)`
