+++ front/src/modules/ui/input/components/DateInputDisplay.tsx
