+++ front/src/modules/ui/input/components/EmailInputDisplay.tsx
