+++ front/src/modules/ui/input/components/__stories__/AutosizeTextInput.stories.tsx
