+++ front/src/modules/ui/input/components/__stories__/Checkbox.stories.tsx
