+++ front/src/modules/ui/input/components/__stories__/EmailInputDisplay.stories.tsx
