+++ front/src/modules/ui/menu-item/components/MenuItemMultiSelectAvatar.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import styled from '@emotion/styled';
 
-import { Checkbox } from '@/ui/input/checkbox/components/Checkbox';
+import { Checkbox } from '@/ui/input/components/Checkbox';
 
 import {
   StyledMenuItemBase,
