+++ front/src/modules/ui/menu-item/components/MenuItemToggle.tsx
@@ -1,5 +1,5 @@
 import { IconComponent } from '@/ui/icon/types/IconComponent';
-import { Toggle } from '@/ui/input/toggle/components/Toggle';
+import { Toggle } from '@/ui/input/components/Toggle';
 
 import { MenuItemLeftContent } from '../internals/components/MenuItemLeftContent';
 import { StyledMenuItemBase } from '../internals/components/StyledMenuItemBase';
